production/company_policy_bot/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
slack-bolt==1.23.0
pypdf==5.4.0
rank-bm25==0.2.2
google-generativeai==0.7.1
google-cloud-aiplatform
google-genai>=0.1.0
faster-whisper==1.0.3
setuptools
nvidia-cublas-cu12
nvidia-cudnn-cu12
